requirements.txt
flask==3.1.0
python-dotenv==1.0.1
rich==13.9.4
azure-identity==1.19.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.0
azure-mgmt-network==27.0.0
